+++ NEWS.md
@@ -9,6 +9,7 @@ Fixes:
   * Gracefully stop in the presence of "" array names (issue [#95](https://github.com/hugomflavio/actel/issues/95)).
   * Fix minor aesthetic bug that would cause invalid detections to show up as NA in the captions of individual detection plots.
   * Fix crash when running migration analysis with only one array (issue [#113](https://github.com/hugomflavio/actel/issues/113)).
+  * Fix crash when running residency analysis and a group of animals is never detected (issue [#114](https://github.com/hugomflavio/actel/issues/114)).
 
 Enhancements:
   * Improve timestamp handling when importing data through `preload()` (issue [#94](https://github.com/hugomflavio/actel/issues/94)).